user functions

+++ SolutionReservation.API/Controllers/UserController.cs
@@ -3,6 +3,7 @@ using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using SolutionReservation.API.DTO.Input;
 using SolutionReservation.API.MapperDTO;
+using SolutionReservation.Data.Model;
 using SolutionReservation.Domain.Managers;
 using SolutionReservation.Domain.Model;
 
@@ -41,6 +42,7 @@ namespace SolutionReservation.API.Controllers
         {
             try
             {
+                if (!await _userManager.ExistsAsync(clientNumber)) return NotFound($"User with ID {clientNumber} not found");
                 var result = await _userManager.GetAsync(clientNumber);
                 return Ok(result);
             }
@@ -56,6 +58,9 @@ namespace SolutionReservation.API.Controllers
         {
             try
             {
+                if (!await _userManager.ExistsAsync(clientNumber)) return NotFound($"User with ID {clientNumber} not found");
+                var userFromDb = await _userManager.GetAsync(clientNumber);
+                user = UserMapperDTO.UpdateUser(userFromDb,user);
                 var result = await _userManager.UpdateUserAsync(clientNumber, UserMapperDTO.ToDomain(user));
                 return Ok(result);
             }
@@ -71,7 +76,40 @@ namespace SolutionReservation.API.Controllers
         {
             try
             {
+                if (!await _userManager.ExistsAsync(clientNumber)) return NotFound($"User with ID {clientNumber} not found");
                 var result = await _userManager.DeleteUserAsync(clientNumber);
+                return Ok(result);
+            }
+            catch (Exception ex)
+            {
+                return StatusCode(StatusCodes.Status500InternalServerError, ex.Message);
+            }
+        }
+
+        [HttpGet]
+        [Route("SearchRestaurant/{search}")]
+        public async Task<IActionResult> SearchRestaurantAsync(string search)
+        {
+            try
+            {
+                var result = await _userManager.SearchRestaurantAsync(search);
+                return Ok(result);
+            }
+            catch (Exception ex)
+            {
+                return StatusCode(StatusCodes.Status500InternalServerError, ex.Message);
+            }
+        }
+
+        [HttpPut]
+        [Route("AddReservation/{clientNumber}/{restaurantId}")]
+        public async Task<IActionResult> AddReservationAsync(int clientNumber, int restaurantId, ReservationInputDTO reservation)
+        {
+            try
+            {
+                if (!await _userManager.ExistsAsync(clientNumber)) return NotFound($"User with ID {clientNumber} not found");
+                if (!await _userManager.ExistsRestaurantAsync(restaurantId)) return NotFound($"Restaurant with ID {restaurantId} not found");
+                var result = await _userManager.AddReservationAsync(clientNumber, restaurantId,ReservationMapperDTO.ToDomain(reservation));
                 return Ok(result);
             }
             catch (Exception ex)
